aarch64 fix compiler complaints

+++ src/platform/aarch64_vm/init_libc.cpp
@@ -38,7 +38,7 @@ int kernel_main(int, char * *, char * *)
   LL_ASSERT(global_ctors_ok == 42);
   kernel::state().libc_initialized = true;
 
-  Elf_binary<Elf64> elf{{(char*)&_ELF_START_, &_ELF_END_ - &_ELF_START_}};
+  Elf_binary<Elf64> elf{{(char*)&_ELF_START_, static_cast<unsigned long>(&_ELF_END_ - &_ELF_START_)}};
   LL_ASSERT(elf.is_ELF() && "ELF header intact");
 
   PRATTLE("<kernel_main> OS start \n");
@@ -85,7 +85,7 @@ namespace aarch64
     auto* ehdr = (Elf64_Ehdr*)&_ELF_START_;
     auto* phdr = (Elf64_Phdr*)((char*)ehdr + ehdr->e_phoff);
     LL_ASSERT(phdr);
-    Elf_binary<Elf64> elf{{(char*)&_ELF_START_, &_ELF_END_ - &_ELF_START_}};
+    Elf_binary<Elf64> elf{{(char*)&_ELF_START_, static_cast<unsigned long>(&_ELF_END_ - &_ELF_START_)}};
     LL_ASSERT(elf.is_ELF());
     LL_ASSERT(phdr[0].p_type == PT_LOAD);
